% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMMIf.R
\name{AMMIf}
\alias{AMMIf}
\title{AMMI (Additive Main effects and Multiplicative Interaction)}
\usage{
AMMIf()
}
\arguments{
\item{df}{dataframe object}

\item{trials}{factor type column associated to environment levels within df object}

\item{gen}{factor type column associated to genotype levels within df object}

\item{rep}{factor type column associated to replication or blocks levels within df object}

\item{y}{numeric type column associated to evaluated trait within df object}

\item{GE}{TRUE or FALSE, return a plot of the predominant type of GxE interaction, genotypic correaltion across trials and variety connectivity}
}
\description{
Returns a dataframe containing outputs results from two-by-two analysis using
mixed model REML/BLUP (assuming random genotypic effects and fixed block).
Perform a analysis on the variety connectivity (number of the same genotypes among trials),
calculate the
Indicates what type of genoytpe x environment interaction are predominant.
}
\examples{
data(MET_maize)
env.corr<-envcorrelation(y = "YIELD", trials = "ENV",
                         gen = "GEN", rep = "REP", df = MET.maize)
head(env.corr)
}
\references{
Eberhart S.A., Russell W.A. (1966) Stability parameters for comparing varieties. Crop Sci. 6: 36-40.

Singh R.K., Chaudhary B.D.(1985) Biometrical Methods in Quantitative Genetics Analysis, Kalyani Publishers
}
\author{
Adapted from Umesh R Rosyara by Germano Costa Neto
}
