% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dBLUPPC.R
\name{DBLUPC}
\alias{DBLUPC}
\title{Correcting phenotypic datasets for GS and GWAS using deregressed BLUPs}
\usage{
DBLUPC(gen, env, rep, traits, df)
}
\arguments{
\item{gen}{factor, column associated to genotype levels within df object}

\item{rep}{factor, column associated to replication or blocks levels within df object}

\item{traits}{character, list of columns associated to evaluated trait within df object}

\item{df}{dataframe object}

\item{trials}{factor, column associated to environment levels within df object}

\item{accuracy}{numeric, a single value (from 0 to 1) indicating the expected threshold accuracy (we recommend the use of 0.7)}
}
\description{
Perform a mixed model analysis of each individual trial within a MET dataset.
}
\examples{
data(MET_maize)
}
\references{
1. Colombari Filho JM, de Resende MDV, de Morais OP, de Castro AP, Guimarães ÉP, Pereira JA, et al. Upland rice breeding in Brazil: A simultaneous genotypic evaluation of stability, adaptability and grain yield. Euphytica. 2013;192(1):117–29.

2. Resende MDV, Duarte JB. Precisão e controle de qualidade em experimentos de avaliação de cultivares. Pesquisa Agropecuária Tropical. 2007;37(3):182–94.
}
\author{
Germano Martins F. Costa Neto  <germano.cneto@usp.br>
}
