% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envcorrelation.R
\name{ENVCOR}
\alias{ENVCOR}
\title{Genotypic Correlation across Environments}
\usage{
ENVCOR(df, trials, gen, rep, y, plot = TRUE)
}
\arguments{
\item{df}{dataframe object}

\item{trials}{factor type column associated to environment levels within df object}

\item{gen}{factor type column associated to genotype levels within df object}

\item{rep}{factor type column associated to replication or blocks levels within df object}

\item{y}{numeric type column associated to evaluated trait within df object}

\item{GE}{TRUE or FALSE, return a plot of the predominant type of GxE interaction, genotypic correaltion across trials and variety connectivity}
}
\description{
Returns a dataframe containing outputs results from two-by-two analysis using
mixed model REML/BLUP (assuming random genotypic effects and fixed block).
Perform a analysis on the variety connectivity (number of the same genotypes among trials),
calculate the
Indicates what type of genoytpe x environment interaction are predominant.
}
\examples{
data(MET_maize)
env.corr<-envcorrelation(y = "YIELD", trials = "ENV",
                         gen = "GEN", rep = "REP", df = MET.maize)
head(env.corr)
}
\references{
1. Colombari Filho JM, de Resende MDV, de Morais OP, de Castro AP, Guimarães ÉP, Pereira JA, et al. Upland rice breeding in Brazil: A simultaneous genotypic evaluation of stability, adaptability and grain yield. Euphytica. 2013;192(1):117–29.

2. Smith AB, Ganesalingam A, Kuchel H, Cullis BR. Factor analytic mixed models for the provision of grower information from national crop variety testing programs. Theor Appl Genet. 2014;128(1):55–72.
}
\author{
Germano Martins F. Costa Neto  <germano.cneto@usp.br>
}
