% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FR.r
\name{FR}
\alias{FR}
\title{Factorial Regression (FR) using additional information}
\usage{
FR(df.y, df.cov, scale = FALSE)
}
\arguments{
\item{df.y}{dataframe contaning response values (e.g., GE matrix)}

\item{df.cov}{dataframe contaning predictor values (e.g., covariate matrix)}

\item{scale}{FALSE (default) or TRUE if scale df.y is required}
}
\description{
Returns a dataframe containing outputs results from two-by-two analysis using
mixed model REML/BLUP (assuming random genotypic effects and fixed block).
Perform a analysis on the variety connectivity (number of the same genotypes among trials),
calculate the
Indicates what type of genoytpe x environment interaction are predominant.
}
\examples{
data(MET_maize)
env.corr<-envcorrelation(y = "YIELD", trials = "ENV",
                         gen = "GEN", rep = "REP", df = MET.maize)
head(env.corr)
}
\references{
Van Eeuwijk FA. Linear and bilinear models for the analysis of multi-environment trials: I. An inventory of models. Euphytica. 1995;84(1):1–7.

Brancourt-Hulmel M, Denis JB, Lecomte C. Determining environmental covariates which explain genotype environment interaction in winter wheat through probe genotypes and biadditive factorial regression. Theor Appl Genet. 2000;100(2):285–98.

Balfourier F, Oliveira JA, Charmet G, Arbones E. Factorial regression analysis of genotype by environment interaction in ryegrass populations, using both isozyme and climatic data as covariates. Euphytica. 1997;98(1):37–46.

Baril CP, Denis J-B, Wustman R, Van Eeuwijk FA. Analysing genotype by environment interaction in Dutch potato variety trials using factorial regression. Euphytica. 1995;84(1):23–9.

Vargas M, Crossa J, Van Eeuwijk FA, Ramírez ME, Sayre K. Using partial least squares regression, factorial regression, and AMMI models for interpreting genotype x environment interaction. Crop Sci. 1999;39(4):955–67.

Costa-Neto GMF. Integrating environmental covariates and thematic maps into genotype by environment interaction analysis in upland rice. Master degree Thesis in Genetics and Plant Breeding, Agronomy School, Federal University of Goiás. Brazil, 2017. 122f.
}
\author{
Germano Costa Neto
}
