% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecovalence.R
\name{GGAW}
\alias{GGAW}
\title{Fixed effects of G, GE, G+GE and Ecovalence}
\usage{
GGAW(df, digits, plot = TRUE)
}
\arguments{
\item{df}{dataframe object}

\item{digits}{numeric, digits for round ouputs}
}
\description{
Returns a dataframe containing outputs results from two-by-two analysis using
mixed model REML/BLUP (assuming random genotypic effects and fixed block).
Perform a analysis on the variety connectivity (number of the same genotypes among trials),
calculate the
Indicates what type of genoytpe x environment interaction are predominant.
}
\examples{
data(corsten.interaction)
m1 <- melt(corsten.interaction, measure.var='yield')
dmat <- acast(m1, loc~gen)
dmat<-ecovalence(as.matrix(dmat),digits=3)
dmat$Ecovalence
dmat$GA.matrix
dmat$Y.j
}
\references{
1. Colombari Filho JM, de Resende MDV, de Morais OP, de Castro AP, Guimarães ÉP, Pereira JA, et al. Upland rice breeding in Brazil: A simultaneous genotypic evaluation of stability, adaptability and grain yield. Euphytica. 2013;192(1):117–29.

2. Smith AB, Ganesalingam A, Kuchel H, Cullis BR. Factor analytic mixed models for the provision of grower information from national crop variety testing programs. Theor Appl Genet. 2014;128(1):55–72.
}
\author{
Germano Martins F. Costa Neto  <germano.cneto@usp.br>
}
