% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrialEvalut.R
\name{TE1}
\alias{TE1}
\title{Mixed model analysis of individuals trials}
\usage{
TE1(df, y, rep, gen, trials, accuracy)
}
\arguments{
\item{df}{dataframe object}

\item{y}{numeric, column associated to evaluated trait within df object}

\item{rep}{factor, column associated to replication or blocks levels within df object}

\item{gen}{factor, column associated to genotype levels within df object}

\item{trials}{factor, column associated to environment levels within df object}

\item{accuracy}{numeric, a single value (from 0 to 1) indicating the expected threshold accuracy (we recommend the use of 0.7)}
}
\description{
Perform a mixed model analysis of each individual trial within a MET dataset.
}
\examples{
data(MET_maize)
trial.evalut-trialEvalut(y = "YIELD", trials = "ENV", accuracy=0.7,
                         gen = "GEN", rep = "REP", df = MET.maize)
head(trial.evalut)
}
\references{
1. Colombari Filho JM, de Resende MDV, de Morais OP, de Castro AP, Guimarães ÉP, Pereira JA, et al. Upland rice breeding in Brazil: A simultaneous genotypic evaluation of stability, adaptability and grain yield. Euphytica. 2013;192(1):117–29.

2. Resende MDV, Duarte JB. Precisão e controle de qualidade em experimentos de avaliação de cultivares. Pesquisa Agropecuária Tropical. 2007;37(3):182–94.
}
\author{
Germano Martins F. Costa Neto  <germano.cneto@usp.br>
}
